Implemented database for level manager script
Using Xampp as a server host and mySQL as the database, i created a small database that holds level id's and level names. The level manager will get the id from a level if their names are the same. When it changes the level, it (currently) increments the level count and loads a new scene based on level name.